+++ test/examples/flappy_bird/browser-entry.text-frame.utils.ts
@@ -18,7 +18,7 @@ import {
   FLAPPY_TITLE_BOX_MIN_OUTER_GAP_COLUMNS,
   FLAPPY_TITLE_BOX_MIN_WIDTH,
   FLAPPY_TITLE_BOX_WIDTH_RATIO,
-} from './browser-entry.constants';
+} from './constants';
 import type {
   RenderClosedOuterBoxInput,
   RenderStandaloneTitleBoxInput,
