+++ test/examples/flappy_bird/browser-entry.viewport.utils.ts
@@ -1,7 +1,4 @@
-import {
-  FLAPPY_HALF,
-  FLAPPY_NORMALIZATION_EPSILON,
-} from './browser-entry.constants';
+import { FLAPPY_HALF, FLAPPY_NORMALIZATION_EPSILON } from './constants';
 import { FLAPPY_WORLD_HEIGHT_PX, FLAPPY_PIPE_WIDTH_PX } from './constants';
 import type { ViewportInfo } from './browser-entry.types';
